refactor visit_FormattedValue, visit_JoinedStr
The idea here is that children nodes should be responsible for their own
rendering, even if it is dependent on their parents. This is consistent
with design choices elsewhere in ast_decompiler. The change also means
that self.node_stack reflects the truth, which is what caused the bug in
a5bd70e.