f-string changes (#6)
* refactor visit_FormattedValue, visit_JoinedStr

The idea here is that children nodes should be responsible for their own
rendering, even if it is dependent on their parents. This is consistent
with design choices elsewhere in ast_decompiler. The change also means
that self.node_stack reflects the truth, which is what caused the bug in
a5bd70e.

* raise NotImplementedError on nested f-strings

Since there is no plan to fix this in the near future (see issues #5),
rather than fail mysteriously, we should be explicit about our (lack of)
intention.

* fix f-string on {}-like objects

Add space so that we don't accidentally escape curly braces. Additional
test cases added.